fixed the WalletResourcesNetworkServicePluginRoot startTest and added the @Ignore annotation to the other tests while Natalia fixes them
