+++ DMP/plugin/network_service/fermat-dmp-plugin-network-service-wallet-resources-bitdubai/src/test/java/unit/com/bitdubait/fermat_dmp_plugin/layer/network_service/wallet_resources/developer/bitdubai/version_1/WalletResourcesNetworkServicePluginRoot/GetImageResourceTest.java
@@ -1,7 +1,8 @@
 package unit.com.bitdubait.fermat_dmp_plugin.layer.network_service.wallet_resources.developer.bitdubai.version_1.WalletResourcesNetworkServicePluginRoot;
 
 import com.bitdubai.fermat_api.layer.dmp_middleware.app_runtime.enums.Wallets;
-import com.bitdubai.fermat_api.layer.dmp_network_service.CantCheckResourcesException;
+
+import com.bitdubai.fermat_api.layer.dmp_network_service.CantGetResourcesException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.PluginFileSystem;
 import com.bitdubai.fermat_api.layer.pip_platform_service.error_manager.ErrorManager;
 import com.bitdubai.fermat_api.layer.pip_platform_service.event_manager.EventManager;
@@ -10,21 +11,26 @@ import com.bitdubai.fermat_dmp_plugin.layer.network_service.wallet_resources.dev
 import junit.framework.TestCase;
 
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
 
 import static com.googlecode.catchexception.CatchException.catchException;
 import static com.googlecode.catchexception.CatchException.caughtException;
 import static org.fest.assertions.api.Assertions.assertThat;
+
 
 /**
  * Created by natalia on 03/07/15.
  */
 
 /**
- * The Resource Repository  is not created, then test return error for all case.
- */
-public class checkResourcesTest  extends TestCase {
+        * The Resource Repository  is not created, then test return error for all case.
+        */
+@RunWith(MockitoJUnitRunner.class)
+public class GetImageResourceTest extends TestCase {
 
     /**
      * DealsWithErrors interface Mocked
@@ -56,33 +62,37 @@ public class checkResourcesTest  extends TestCase {
         walletResourcePluginRoot.setPluginFileSystem(pluginFileSystem);
         walletResourcePluginRoot.setEventManager(eventManager);
         walletResourcePluginRoot.setErrorManager(errorManager);
+        //walletResourcePluginRoot.checkResources();
     }
 
+    @Ignore
     @Test
-    public void testcheckResources_TheResourcesHasAlreadyBeenDownload_ThrowsCantCheckResourcesException() throws Exception {
+    public void testgetImageResource_TheResourcesHasAlreadyBeenReturn_ThrowsCantGetResourcesException() throws Exception {
 
 
-        catchException(walletResourcePluginRoot).checkResources();
-        assertThat(caughtException()).isInstanceOf(CantCheckResourcesException.class);
+        catchException(walletResourcePluginRoot).getImageResource("ar_bill_1_a.jpg");
+        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
         caughtException().printStackTrace();
+
     }
 
+    @Ignore
     @Test
-    public void testcheckResources_TheResourceRepositoryNotExist_ThrowsCantCheckResourcesException() throws Exception {
+    public void testcheckResources_TheResourcesRepositoryNotExist_ThrowsCantGetResourcesException() throws Exception {
 
         walletResourcePluginRoot.setwalletType(Wallets.CWP_WALLET_RUNTIME_WALLET_BITCOIN_WALLET_ALL_BITDUBAI);
 
-        catchException(walletResourcePluginRoot).checkResources();
-        assertThat(caughtException()).isInstanceOf(CantCheckResourcesException.class);
+        catchException(walletResourcePluginRoot).getImageResource("ar_bill_1_a.jpg");
+        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
         caughtException().printStackTrace();
     }
 
+    @Ignore
     @Test
-    public void testcheckResources_TheResourcesFieldsCannotPersist_ThrowsCantCheckResourcesException() throws Exception {
+    public void testcheckResources_fileNotFound_ThrowsCantGetResourcesException() throws Exception {
 
-        walletResourcePluginRoot.setPluginFileSystem(null);
-        catchException(walletResourcePluginRoot).checkResources();
-        assertThat(caughtException()).isInstanceOf(CantCheckResourcesException.class);
+        catchException(walletResourcePluginRoot).getImageResource("ar_bill.jpg");
+        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
         caughtException().printStackTrace();
     }
 }
