+++ DMP/plugin/network_service/fermat-dmp-plugin-network-service-wallet-resources-bitdubai/src/test/java/unit/com/bitdubait/fermat_dmp_plugin/layer/network_service/wallet_resources/developer/bitdubai/version_1/WalletResourcesNetworkServicePluginRoot/GetLayoutResourceTest.java
@@ -1,8 +1,7 @@
 package unit.com.bitdubait.fermat_dmp_plugin.layer.network_service.wallet_resources.developer.bitdubai.version_1.WalletResourcesNetworkServicePluginRoot;
 
-import com.bitdubai.fermat_api.CantStartPluginException;
 import com.bitdubai.fermat_api.layer.dmp_middleware.app_runtime.enums.Wallets;
-
+import com.bitdubai.fermat_api.layer.dmp_network_service.CantGetResourcesException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.PluginFileSystem;
 import com.bitdubai.fermat_api.layer.pip_platform_service.error_manager.ErrorManager;
 import com.bitdubai.fermat_api.layer.pip_platform_service.event_manager.EventManager;
@@ -11,8 +10,11 @@ import com.bitdubai.fermat_dmp_plugin.layer.network_service.wallet_resources.dev
 import junit.framework.TestCase;
 
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
 
 import static com.googlecode.catchexception.CatchException.catchException;
 import static com.googlecode.catchexception.CatchException.caughtException;
@@ -21,7 +23,12 @@ import static org.fest.assertions.api.Assertions.assertThat;
 /**
  * Created by natalia on 03/07/15.
  */
-public class startTest extends TestCase {
+
+/**
+ * The Resource Repository  is not created, then test return error for all case.
+ */
+@RunWith(MockitoJUnitRunner.class)
+public class GetLayoutResourceTest extends TestCase {
 
     /**
      * DealsWithErrors interface Mocked
@@ -53,15 +60,40 @@ public class startTest extends TestCase {
         walletResourcePluginRoot.setPluginFileSystem(pluginFileSystem);
         walletResourcePluginRoot.setEventManager(eventManager);
         walletResourcePluginRoot.setErrorManager(errorManager);
+
+        //walletResourcePluginRoot.checkResources();
     }
 
+    @Ignore
     @Test
-    public void teststart_ThePlugInHasStartedOk_ThrowsCantStartPluginException() throws Exception {
+    public void testgetImageResource_TheResourcesHasAlreadyBeenReturn_ThrowsCantGetResourcesException() throws Exception {
 
 
-        catchException(walletResourcePluginRoot).start();
-        assertThat(caughtException()).isInstanceOf(NullPointerException.class);
+        catchException(walletResourcePluginRoot).getLayoutResource("wallets_kids_fragment_balance.txt");
+        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
+        caughtException().printStackTrace();
+
+    }
+
+    @Ignore
+    @Test
+    public void testcheckResources_TheResourcesRepositoryNotExist_ThrowsCantGetResourcesException() throws Exception {
+
+        walletResourcePluginRoot.setwalletType(Wallets.CWP_WALLET_RUNTIME_WALLET_BITCOIN_WALLET_ALL_BITDUBAI);
+
+        catchException(walletResourcePluginRoot).getLayoutResource("wallets_kids_fragment_balance.txt");
+        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
         caughtException().printStackTrace();
     }
 
+    @Ignore
+    @Test
+    public void testcheckResources_fileNotFound_ThrowsCantGetResourcesException() throws Exception {
+
+        catchException(walletResourcePluginRoot).getLayoutResource("layout1.xml");
+        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
+        caughtException().printStackTrace();
+    }
 }
+
+
