--- DMP/plugin/network_service/fermat-dmp-plugin-network-service-wallet-resources-bitdubai/src/test/java/unit/com/bitdubait/fermat_dmp_plugin/layer/network_service/wallet_resources/developer/bitdubai/version_1/WalletResourcesNetworkServicePluginRoot/getImageResourceTest.java
@@ -1,91 +1,0 @@
-package unit.com.bitdubait.fermat_dmp_plugin.layer.network_service.wallet_resources.developer.bitdubai.version_1.WalletResourcesNetworkServicePluginRoot;
-
-import com.bitdubai.fermat_api.layer.dmp_middleware.app_runtime.enums.Wallets;
-
-import com.bitdubai.fermat_api.layer.dmp_network_service.CantGetResourcesException;
-import com.bitdubai.fermat_api.layer.osa_android.file_system.PluginFileSystem;
-import com.bitdubai.fermat_api.layer.pip_platform_service.error_manager.ErrorManager;
-import com.bitdubai.fermat_api.layer.pip_platform_service.event_manager.EventManager;
-import com.bitdubai.fermat_dmp_plugin.layer.network_service.wallet_resources.developer.bitdubai.version_1.WalletResourcesNetworkServicePluginRoot;
-
-import junit.framework.TestCase;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.Mock;
-
-import static com.googlecode.catchexception.CatchException.catchException;
-import static com.googlecode.catchexception.CatchException.caughtException;
-import static org.fest.assertions.api.Assertions.assertThat;
-
-
-/**
- * Created by natalia on 03/07/15.
- */
-
-/**
-        * The Resource Repository  is not created, then test return error for all case.
-        */
-public class getImageResourceTest extends TestCase {
-
-    /**
-     * DealsWithErrors interface Mocked
-     */
-    @Mock
-    ErrorManager errorManager;
-
-    /**
-     * UsesFileSystem Interface member variables.
-     */
-    @Mock
-    PluginFileSystem pluginFileSystem;
-
-
-    /**
-     * DealWithEvents Iianterface member variables.
-     */
-    @Mock
-    EventManager eventManager;
-
-
-
-    WalletResourcesNetworkServicePluginRoot walletResourcePluginRoot;
-
-    @Before
-    public void setUp() throws Exception {
-        walletResourcePluginRoot = new WalletResourcesNetworkServicePluginRoot();
-        walletResourcePluginRoot.setwalletType(Wallets.CWP_WALLET_RUNTIME_WALLET_AGE_KIDS_ALL_BITDUBAI);
-        walletResourcePluginRoot.setPluginFileSystem(pluginFileSystem);
-        walletResourcePluginRoot.setEventManager(eventManager);
-        walletResourcePluginRoot.setErrorManager(errorManager);
-        //walletResourcePluginRoot.checkResources();
-    }
-
-    @Test
-    public void testgetImageResource_TheResourcesHasAlreadyBeenReturn_ThrowsCantGetResourcesException() throws Exception {
-
-
-        catchException(walletResourcePluginRoot).getImageResource("ar_bill_1_a.jpg");
-        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
-        caughtException().printStackTrace();
-
-    }
-
-    @Test
-    public void testcheckResources_TheResourcesRepositoryNotExist_ThrowsCantGetResourcesException() throws Exception {
-
-        walletResourcePluginRoot.setwalletType(Wallets.CWP_WALLET_RUNTIME_WALLET_BITCOIN_WALLET_ALL_BITDUBAI);
-
-        catchException(walletResourcePluginRoot).getImageResource("ar_bill_1_a.jpg");
-        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
-        caughtException().printStackTrace();
-    }
-
-    @Test
-    public void testcheckResources_fileNotFound_ThrowsCantGetResourcesException() throws Exception {
-
-        catchException(walletResourcePluginRoot).getImageResource("ar_bill.jpg");
-        assertThat(caughtException()).isInstanceOf(CantGetResourcesException.class);
-        caughtException().printStackTrace();
-    }
-}
